+++ src/Base/Foundation/Plugin.php
@@ -25,7 +25,7 @@ class Plugin
      *
      * @var string VERSION
      */
-    const VERSION = '3.2.5';
+    const VERSION = '3.2.6';
 
     /**
      * Path to the root of the plugin.
@@ -87,12 +87,6 @@ class Plugin
 
         // Set up service providers
         $this->callServiceProviders('register');
-
-        if (\is_admin()) {
-            $this->callServiceProviders('register', 'admin');
-            $this->callServiceProviders('boot', 'admin');
-        }
-
         $this->callServiceProviders('boot');
 
         // Register the Hook loader.
@@ -122,10 +116,9 @@ class Plugin
      *
      * @throws \Exception
      */
-    public function callServiceProviders($method, $key = '')
+    public function callServiceProviders($method)
     {
-        $offset   = $key ? "core.providers.{$key}" : 'core.providers';
-        $services = $this->config->get($offset);
+        $services = $this->config->get('core.providers');
 
         foreach ($services as $service) {
             if (is_array($service)) {